Rename dangling line to boundary line (#3624)
* update : rename dangling line to boundary line
* Serialization of criteria, actions and limit reductions
Simplification of compatibility serde for dangling lines
* Review rework
* upgrade limit reduction version

+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/json/limitreduction/LimitReductionDeserializer.java
@@ -30,7 +30,7 @@ public class LimitReductionDeserializer extends StdDeserializer<LimitReduction> 
     }
 
     private static final class ParsingContext {
-        float value;
+        double value;
         LimitType limitType;
         Boolean monitoringOnly;
         ContingencyContext contingencyContext;
@@ -45,7 +45,7 @@ public class LimitReductionDeserializer extends StdDeserializer<LimitReduction> 
             switch (fieldName) {
                 case "value" -> {
                     parser.nextToken();
-                    context.value = parser.readValueAs(Float.class);
+                    context.value = parser.readValueAs(Double.class);
                     return true;
                 }
                 case "limitType" -> {
@@ -77,7 +77,7 @@ public class LimitReductionDeserializer extends StdDeserializer<LimitReduction> 
             }
         });
         LimitReduction.Builder builder = LimitReduction.builder(checkAttribute(context.limitType, "limitType"),
-                        checkAttribute(context.value, "value"))
+                checkAttribute(context.value, "value"))
                 .withMonitoringOnly(checkAttribute(context.monitoringOnly, "monitoringOnly"))
                 .withContingencyContext(checkAttribute(context.contingencyContext, "contingencyContext"));
         if (context.networkElementCriteria != null) {
